app/requirements.txt
wheel
cassandra-driver
venv-pack
pyarrow
fastparquet
pathvalidate
pyspark
tqdm
nltk
cqlsh